test: align launch attestation remediation coverage
